Update the test file name
